Устранение ошибки при которой объект активного бота не создавался в бд. Удаление лишних принтов системной информации. Очистка бд в блоке файнали работы бота.

+++ traider_bot/bots/bot_logic.py
@@ -576,7 +576,8 @@ def lock_release():
 
 def exit_by_exception(bot):
     bot_id = bot.pk
-    ActiveBot.objects.filter(bot_id=bot_id).delete()
+    if ActiveBot.objects.filter(bot_id=bot_id):
+        ActiveBot.objects.filter(bot_id=bot_id).delete()
 
     if bot_id in global_list_bot_id:
         global_list_bot_id.remove(bot_id)
